Crate hexgrid

Hexagonal Grid Map Utility library

A lot of ideas taken from redbloggames hexagon page

Pointy-topped:

          /\
        /    \
       |      |
       |      |
        \    /
          \/

           -z
+y     YZ  |  XZ     +x
 ---       |       ---
    ---    |    ---
       --- | ---
  YX      -x-    XY
       --- | ---
    ---    |    ---
 ---   ZX  |  ZY   ---
-x         |          -y
           +z

Flat-topped:

           ____
          /    \
         /      \
         \      /
          \____/

       +y       -z
        \       /
         \ YZ  /
      YX  \   /  XZ
           \ /
  -x--------x--------+x
           / \
      ZX  /   \ XY
         /  ZY \
        /       \
       +z       -y

Structs

Bearing

Bearing on 2d hexagonal grid (Coordinate + Direction)

Coordinate

Integer Coordinate on 2d hexagonal grid

LineToIter

An iterator over an a line of Coordinates

LineToLossyIter

An iterator over an a line of Coordinates, using a lossy algorithm

LineToWithEdgeDetection

An iterator over an a line of Coordinates, with edge detection

Offset

Coordinate Offset

Position

Fractional Position on 2d hexagonal grid

RangeIter

Iterator over an range

Ring

Iterator over a ring

Spiral

Iterates through all hexes in a hexagonal shaped spiral pattern.

Enums

Angle

Angle, relative to a Direction

Direction

Direction on a hexagonal map

IntegerSpacing

Integer pixel tile size for integer pixel conversion functions

Spacing

Floating point tile size for pixel conversion functions

Spin

Spinning directions

Traits

Float

Float trait required by this library

Integer

Integer trait required by this library