Struct hexgrid::Coordinate

pub struct Coordinate<I: Integer = i32> {
    pub x: I,
    pub y: I,
}

Integer Coordinate on 2d hexagonal grid

          /\
        /    \
       |      |
       |      |
        \    /
          \/

           -z
+y     YZ  |  XZ     +x
 ---       |       ---
    ---    |    ---
       --- | ---
  YX      -x-    XY
       --- | ---
    ---    |    ---
 ---   ZX  |  ZY   ---
-x         |          -y
           +z

Flat-topped:

           ____
          /    \
         /      \
         \      /
          \____/

       +y       -z
        \       /
         \ YZ  /
      YX  \   /  XZ
           \ /
  -x--------x--------+x
           / \
      ZX  /   \ XY
         /  ZY \
        /       \
       +z       -y

Fields
x: I

x coordinate

y: I

y coordinate

Implementations

impl<I: Integer> Coordinate<I>

pub fn from_cubic(x: I, y: I) -> Coordinate<I>

Create new Coord from x and y

pub fn position<F: Float>(&self) -> Position<F>

Convert into fractional position.

pub fn nearest<F: Float>(x: F, y: F) -> Coordinate<I>

Round x, y float to nearest hex coordinates

pub fn nearest_lossy<F: Float>(x: F, y: F) -> Option<Coordinate<I>>

Round x, y float to nearest hex coordinates

Return None, if exactly on the border of two hex coordinates

pub fn from_pixel<F: Float>(x: F, y: F, spacing: Spacing<F>) -> Coordinate<I>

Find the hex containing a pixel. The origin of the pixel coordinates is the center of the hex at (0,0) in hex coordinates.

pub fn nearest_with_offset(
spacing: IntegerSpacing<I>,
v: (I, I)
) -> (Coordinate<I>, (I, I))

Convert integer pixel coordinates v using spacing to nearest coordinate that has both integer pixel coordinates lower or equal to v. Also return offset (in integer pixels) from that coordinate.

Useful for ASCII visualization. Shamelessly stolen from Hex2d.

pub fn to_cartesian_center<F: Float>(&self, spacing: Spacing<F>) -> (F, F)

Convert to pixel coordinates using spacing, where the parameter means the edge length of a hexagon.

This function is meant for graphical user interfaces where resolution is big enough that floating point calculation make sense.

pub fn to_corner_points<F: Float>(&self, spacing: Spacing<F>) -> [(F, F); 6]

Create a set of ordered points to draw a hex.

Reference: Layout Drawing a Hex

pub fn to_pixel_integer(&self, spacing: IntegerSpacing<I>) -> (I, I)

Convert to integer pixel coordinates using spacing, where the parameters mean the width and height multiplications

pub fn scale(&self, s: I) -> Coordinate<I>

Scale coordinate by a factor s

pub fn neighbors(&self) -> [Coordinate<I>; 6]

Array with all the neighbors of a coordinate

pub fn rotate_around_zero(&self, a: Angle) -> Coordinate<I>

Rotate self around a point (0, 0, 0) using angle of rotation a

pub fn rotate_around(&self, center: Coordinate<I>, a: Angle) -> Coordinate<I>

Rotate self around a center using angle of rotation a

pub fn line_to_iter(&self, dest: Coordinate<I>) -> LineToIter<I>

Iterator over each coordinate in straight line from self to dest

pub fn line_to_lossy_iter(&self, dest: Coordinate<I>) -> LineToLossyIter<I>

Iterator over each coordinate in straight line from self to dest

Skip points on the border of two tiles

pub fn line_to_with_edge_detection_iter(
&self,
dest: Coordinate<I>
) -> LineToWithEdgeDetection<I>

Iterator over each coordinate in straight line from self to dest

On edge condition the pair contains different members, otherwise it’s the same.

pub fn z(&self) -> I

Z coordinate

pub fn direction_from_center_cw(&self) -> Option<Direction>

Direction from center (0, 0) to coordinate

In case of diagonals (edge of two major directions), prefers direction that is clockwise from the diagonal

Returns: None if is center

use hexgrid::{Angle, Direction, Coordinate};

let center = Coordinate::from_cubic(0, 0);

assert_eq!(center.direction_from_center_cw(), None);

for &d in Direction::all() {
    assert_eq!((center + d).direction_from_center_cw(), Some(d));
    assert_eq!((center + d + (d + Angle::LeftForward)).direction_from_center_cw(), Some(d));
    assert_eq!((center + d + (d + Angle::RightForward)).direction_from_center_cw(), Some(d + Angle::RightForward));
}

pub fn directions_from_center(&self) -> Vec<Direction>

Directions that lead from center to a given point.

Returns an array of one or two directions.

pub fn direction_from_center_ccw(&self) -> Option<Direction>

Direction from center (0, 0) to coordinate

In case of diagonals (edge of two major directions), prefers direction that is counter-clockwise from the diagonal.

Returns: None if is center

use hexgrid::{Angle, Direction, Coordinate};

let center = Coordinate::from_cubic(0, 0);

assert_eq!(center.direction_from_center_ccw(), None);

for &d in Direction::all() {
    assert_eq!((center + d).direction_from_center_ccw(), Some(d));
    assert_eq!((center + d + (d + Angle::LeftForward)).direction_from_center_ccw(), Some(d + Angle::LeftForward));
    assert_eq!((center + d + (d + Angle::RightForward)).direction_from_center_ccw(), Some(d));
}

pub fn directions_to(&self, coordinate: Coordinate<I>) -> Vec<Direction>

Directions from self to coordinate

pub fn direction_to_cw(&self, coordinate: Coordinate<I>) -> Option<Direction>

Direction from self to coordinate

In case of diagonals (edge of two major directions), prefers direction that is clockwise from the diagonal.

Returns: None if is center

pub fn direction_to_ccw(&self, coordinate: Coordinate<I>) -> Option<Direction>

Direction from self to coordinate

In case of diagonals (edge of two major directions), prefers direction that is counter-clockwise from the diagonal.

Returns: None if is center

pub fn distance(&self, coordinate: Coordinate<I>) -> I

Distance between two Coordinates

pub fn range_iter(&self, radius: I) -> RangeIter<I>

An iterator over all coordinates in radius r

pub fn ring_iter(&self, radius: I, spin: Spin) -> Ring<I>

Iterator over each coordinate in a ring

Example: Elements in order for Ring of radius 2, Direction ZX, CCW

             8
           9   7
        10   .   6
           .   .
        11   x   5
           .   .
         0   .   4
           1   3
             2

use hexgrid::{Coordinate, Spin, Direction};

let center = Coordinate::from_cubic(5, -1);

for &c in &center.neighbors() {
    for ring_c in c.ring_iter(5, Spin::CCW(Direction::XY)) {
        assert_eq!(c.distance(ring_c), 5);
    }
}

pub fn spiral_iter(&self, radius: I, spin: Spin) -> Spiral<I>

Iterator over each coordinate in a spiral