Struct hexgrid::Position

pub struct Position<F: Float = f32> {
    pub q: F,
    pub r: F,
}

Fractional Position on 2d hexagonal grid

Fields
q: F

Axial q coordinate

r: F

Axial r coordinate

Implementations

impl<F: Float> Position<F>

pub fn from_axial(q: F, r: F) -> Position<F>

Create new Coordinate from q and r axial components.

See Hex Coordinates.

pub fn from_cubic(x: F, y: F, z: F) -> Position<F>

Create new Coordinate from x, y, and z cubic components.

See Hex Coordinates.

pub fn from_coordinate<I: Integer>(coordinate: Coordinate<I>) -> Position<F>

Create Position from Hex Coordinate.

pub fn from_cartesian(x: F, y: F, spacing: Spacing<F>) -> Position<F>

Convert a cartesian point into a hex position. The origin of the cartesian space is the center of the hex at (0,0) in hex coordinates.

See Screen to Hex.

pub fn into_cartesian(&self, spacing: Spacing<F>) -> (F, F)

Convert Position into Cartesian coordinates with their origin at (0, 0).

See Hex to Pixel.

pub fn x(&self) -> F

Cubic Hex X coordinate

pub fn y(&self) -> F

Cubic Hex Y coordinate

pub fn z(&self) -> F

Cubic Hex Z coordinate

pub fn length(&self) -> F

Distance from Origin

pub fn distance(&self, position: Position<F>) -> F

Distance between two Positions in Hex Space

pub fn into_coordinate<I: Integer>(&self) -> Coordinate<I>

Round Position into the nearest Coordinate

See Coordinate::from_position. See Hex Rounding.