Crate hexgrid

Hexagonal Grid Map Utility library

A big thanks to RedBlogGames hexagon page and Hex2d-rs for providing a lot of ideas and some example code I could mutate.

Hex Orientations

Flat Topped:

Pointy Topped:

Structs

Bearing

Coordinate and Directional Component Pair on a 2d hexagonal grid.

Coordinate

Integer Coordinate on 2d hexagonal grid

LineGenIter

Generic iterator over a line returns Positions.

LineIter

An iterator over an a line of Coordinates

LineIterWithEdgeDetection

An iterator over an a line of Coordinates, with edge detection

LossyLineIter

An iterator over an a line of Coordinates, using a lossy algorithm

Offset

Coordinate Offset

Position

Fractional Position on 2d hexagonal grid

RangeIter

Iterator over an range

Ring

Iterates in a ring around an origin at provided radius.

SparseHexStorage

Sparse Hex Grid Data Storage

Spiral

Iterates through all hexes in a hexagonal shaped spiral pattern.

Enums

Angle

Angle, relative to a Direction

Direction

Cubic Direction on a hexagonal map.

IntegerSpacing

Integer pixel tile size for integer pixel conversion functions

Sextant

Which slice of a hexgrid does a coordinate exist in.

Spacing

Floating point tile size for pixel conversion functions

Spin

Spin Direction

Traits

Float

Float trait used by this library

Integer

Integer trait used by this library.