Struct hexgrid::Coordinate

pub struct Coordinate<I: Integer = i32> {
    pub q: I,
    pub r: I,
}

Integer Coordinate on 2d hexagonal grid

Coordinate Components (Cubic -> Axial):

x -> q

y -> -q - r

z -> r

Hex Orientations

Flat Topped:

Pointy Topped:

Fields
q: I

q coordinate

r: I

r coordinate

Implementations

impl<I: Integer> Coordinate<I>

pub fn from_axial(q: I, r: I) -> Coordinate<I>

Create new Coordinate from q and r axial components.

See Hex Coordinates

pub fn from_cubic(x: I, y: I, z: I) -> Coordinate<I>

Create new Coordinate from x, y, and z cubic components.

See Hex Coordinates

pub fn from_position<F: Float>(position: Position<F>) -> Coordinate<I>

Round position to nearest hex coordinates

See Hex Rounding

pub fn from_position_lossy<F: Float>(
position: Position<F>
) -> Option<Coordinate<I>>

Round x, y float to nearest hex coordinates.

Return None, if exactly on the border of two hex coordinates

pub fn x(&self) -> I

Cubic X component.

pub fn y(&self) -> I

Cubic Y component.

pub fn z(&self) -> I

Cubic Z component.

pub fn length(&self) -> I

Distance from Origin

pub fn into_position<F: Float>(&self) -> Position<F>

Convert Coordinate into fractional Position.

pub fn from_cartesian<F: Float>(
x: F,
y: F,
spacing: Spacing<F>
) -> Coordinate<I>

Find the hex containing a pixel. The origin of the pixel coordinates is the center of the hex at (0,0) in hex coordinates.

pub fn nearest_with_offset(
spacing: IntegerSpacing<I>,
v: (I, I)
) -> (Coordinate<I>, (I, I))

Convert integer pixel coordinates v using spacing to nearest coordinate that has both integer pixel coordinates lower or equal to v. Also return offset (in integer pixels) from that coordinate.

Useful for ASCII visualization. Shamelessly stolen from Hex2d.

pub fn cartesian_center<F: Float>(&self, spacing: Spacing<F>) -> (F, F)

Convert center point to cartesian coordinates using spacing, where the size parameter is the edge length of a hexagon.

This function is meant for graphical user interfaces where resolution is big enough that floating point calculation make sense.

See Hex to Screen

pub fn to_corner_points<F: Float>(&self, spacing: Spacing<F>) -> [(F, F); 6]

Create a set of ordered cartesian points to draw a hex.

Reference: Layout Drawing a Hex

pub fn to_pixel_integer(&self, spacing: IntegerSpacing<I>) -> (I, I)

Convert to integer pixel coordinates using spacing, where the parameters mean the width and height multiplications

pub fn scale(&self, scalar: I) -> Coordinate<I>

Scale coordinate by a factor scalar

See Coordinate arithmetic

pub fn neighbors(&self) -> [Coordinate<I>; 6]

Array with all the neighbors of a coordinate

pub fn rotate_around_zero(&self, a: Angle) -> Coordinate<I>

Rotate self around a point (0, 0, 0) using angle of rotation a

pub fn rotate_around(&self, center: Coordinate<I>, a: Angle) -> Coordinate<I>

Rotate self around a center using angle of rotation a

pub fn line_iter(&self, target: Coordinate<I>) -> LineIter<I>

Iterator over each coordinate in straight line from self to dest

pub fn lossy_line_iter(&self, target: Coordinate<I>) -> LossyLineIter<I>

Iterator over each coordinate in straight line from self to dest

Skip points on the border of two tiles

pub fn line_iter_with_edge_detection(
&self,
target: Coordinate<I>
) -> LineIterWithEdgeDetection<I>

Iterator over each coordinate in straight line from self to dest

On edge condition the pair contains different members, otherwise it’s the same.

pub fn direction_from_center_cw(&self) -> Option<Direction>

Direction from center (0, 0) to coordinate

In case of diagonals (edge of two major directions), prefers direction that is clockwise from the diagonal

Returns: None if is center

pub fn directions_from_center(&self) -> Vec<Direction>

Directions that lead from center to a given point.

Returns an array of one or two directions.

pub fn direction_from_center_ccw(&self) -> Option<Direction>

Direction from center (0, 0) to coordinate

In case of diagonals (edge of two major directions), prefers direction that is counter-clockwise from the diagonal.

Returns: None if is center

TODO: Example/Function broken.

use hexgrid::{Angle, Direction, Coordinate};

let center = Coordinate::from_axial(0, 0);

assert_eq!(center.direction_from_center_ccw(), None);

for &d in Direction::all() {
    assert_eq!((center + d).direction_from_center_ccw(), Some(d));
    assert_eq!((center + d + (d + Angle::LeftForward)).direction_from_center_ccw(), Some(d + Angle::LeftForward));
    assert_eq!((center + d + (d + Angle::RightForward)).direction_from_center_ccw(), Some(d));
}

pub fn directions_to(&self, coordinate: Coordinate<I>) -> Vec<Direction>

Directions from self to coordinate

pub fn direction_to_cw(&self, coordinate: Coordinate<I>) -> Option<Direction>

Direction from self to coordinate

In case of diagonals (edge of two major directions), prefers direction that is clockwise from the diagonal.

Returns: None if is center

pub fn direction_to_ccw(&self, coordinate: Coordinate<I>) -> Option<Direction>

Direction from self to coordinate

In case of diagonals (edge of two major directions), prefers direction that is counter-clockwise from the diagonal.

Returns: None if is center

pub fn distance(&self, coordinate: Coordinate<I>) -> I

Distance between two Coordinates

pub fn range_iter(&self, radius: I) -> RangeIter<I>

An iterator over all coordinates in radius r

pub fn ring_iter(&self, radius: I, spin: Spin) -> Ring<I>

Iterates in a ring around an origin at provided radius.

Example: Elements in order for Ring of radius 2, Direction YZ, CW

use hexgrid::{Coordinate, Spin, Direction};

let center = Coordinate::from_axial(5, -1);

for &c in &center.neighbors() {
    for ring_c in c.ring_iter(5, Spin::CCW(Direction::XY)) {
        assert_eq!(c.distance(ring_c), 5);
    }
}

pub fn spiral_iter(&self, radius: I, spin: Spin) -> Spiral<I>

Iterates in a spiral around an origin out to the supplied radius.

Example: Elements in order for all rings from radius 0 to radius 2, Direction ZX, CCW

Ring Iteration Pattern:

use hexgrid::{Coordinate, Spin, Direction};

let locations = Coordinate::from_axial(0, 0).spiral_iter(2, Spin::CW(Direction::YZ)).collect::<Vec<_>>();

assert_eq!(locations.len(), 19);
assert_eq!(locations[0].distance(Coordinate::from_axial(0,0)), 0);
assert_eq!(locations[1].distance(Coordinate::from_axial(0,0)), 1);
assert_eq!(locations[7].distance(Coordinate::from_axial(0,0)), 2);